refactor(client): rename service worker files
