+++ packages/client/src/features/game/sw/startServiceWorker.ts
@@ -1,0 +1,11 @@
+export const startServiceWorker = () => {
+  if ('serviceWorker' in navigator) {
+    window.addEventListener('load', () => {
+      navigator.serviceWorker
+        .register('/cashOnlyServiceWorker.ts')
+        .catch((error: string) => {
+          console.error('Ошибка при регистрации service worker: ', error)
+        })
+    })
+  }
+}
