+++ packages/client/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
 import { App } from './app/app'
 import { ErrorBoundary } from './components/errorBoundary/errorBoundary'
-import { startServiceWorker } from './features/game/sw/serviceWorker'
+import { startServiceWorker } from './features/game/sw/startServiceWorker'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <StrictMode>
